moved logic from cli calls into relevant modules
