+++ vast_post_processing/cli/selavy_combined.py
@@ -7,47 +7,12 @@ to the SLURM queue externally - this script may not have access to the SLURM exe
 from pathlib import Path
 from typing import Optional, List
 
-from loguru import logger
 import typer
 
+from vast_post_processing import combine
+
+
 app = typer.Typer()
-
-
-def write_selavy_files(
-    field_name: str,
-    epoch_name: str,
-    image_path: Path,
-    parset_template_path: Path,
-    sbatch_template_path: Path,
-    weights_path: Optional[Path] = None,
-):
-    if image_path is None:
-        raise FileNotFoundError(f"Image {image_path} doesn't exist.")
-    if weights_path is None:
-        # try to find the weights file using the combined naming convention
-        weights_path = image_path.with_name(f"{image_path.stem}.weight.fits")
-    if not weights_path.exists():
-        raise FileNotFoundError(f"Weights image {weights_path} doesn't exist.")
-
-    image_name = image_path.stem
-    weights_name = weights_path.stem
-
-    parset_template = parset_template_path.read_text().format(
-        image_name=image_name, weights_name=weights_name
-    )
-    parset_path = image_path.with_name(f"selavy.{image_name}.in")
-    parset_path.write_text(parset_template)
-
-    sbatch_template = sbatch_template_path.read_text().format(
-        job_name=f"selavy-{field_name}-{epoch_name}",
-        parset_path=parset_path.relative_to(image_path.parent),
-        log_path=parset_path.with_suffix(".log").relative_to(image_path.parent),
-        working_dir_path=parset_path.parent,
-    )
-    sbatch_path = image_path.with_name(f"selavy.{image_name}.sbatch")
-    sbatch_path.write_text(sbatch_template)
-
-    return sbatch_path
 
 
 @app.command()
@@ -59,25 +24,11 @@ def main(
     racs: bool = False,
     field_list: Optional[List[str]] = typer.Option(None, "--field"),
 ):
-    glob_expr = "RACS_*" if racs else "VAST_*"
-    for field_path in neighbour_data_dir.glob(glob_expr):
-        if field_list and field_path.name not in field_list:
-            logger.info(
-                f"Glob found field {field_path} but it was not given as a --field option."
-                " Skipping."
-            )
-            continue
-        field_name = field_path.name
-        epoch_name = field_path.parent.name
-        image_path = field_path / f"{field_name}.{epoch_name}.{stokes}.conv.fits"
-        try:
-            _ = write_selavy_files(
-                field_name,
-                epoch_name,
-                image_path,
-                parset_template_path,
-                sbatch_template_path,
-            )
-        except FileNotFoundError as e:
-            logger.error(e)
-            continue
+    combine.selavy_combined(
+        neighbour_data_dir,
+        parset_template_path,
+        sbatch_template_path,
+        stokes,
+        racs,
+        field_list,
+    )
