+++ vast_post_processing/utils/fileutils.py
@@ -24,3 +24,11 @@ def cleanup_directory(directory: Path):
             logger.info(f"Deleted directory {path}.")
         else:
             logger.debug(f"Leaving {path}.")
+
+
+def cleanup(neighbour_data_dir: Path, delete_all: bool = False):
+    if delete_all:
+        rmtree(neighbour_data_dir)
+    else:
+        for field_path in neighbour_data_dir.glob("VAST_*"):
+            cleanup_directory(field_path)
